only 1 board impl, destroy piece enum, 'game' abstraction
